fix: tree detachment action
